+++ source/Shared/Graphics/OpenGL1/TextureImplOpenGL1.hpp
@@ -15,10 +15,10 @@ namespace LDL
 		class TextureImplOpenGL1: public TextureImpl
 		{
 		public:
-			TextureImplOpenGL1(RenderContextImpl* renderContextImpl, const LDL::Math::Vec2u & size, uint8_t* pixels, size_t bytesPerPixel);
-			TextureImplOpenGL1(RenderContextImpl* renderContextImpl, const Math::Vec2u& size, size_t bytesPerPixel);
+			TextureImplOpenGL1(RenderContextImpl* renderContextImpl, const LDL::Math::Vec2u & size, uint8_t* pixels, uint8_t bytesPerPixel);
+			TextureImplOpenGL1(RenderContextImpl* renderContextImpl, const Math::Vec2u& size, uint8_t bytesPerPixel);
 			~TextureImplOpenGL1();
-			void Copy(const Math::Vec2u& dstPos, const Math::Vec2u& srcSize, uint8_t * pixels, size_t bytesPerPixel);
+			void Copy(const Math::Vec2u& dstPos, const Math::Vec2u& srcSize, uint8_t * pixels, uint8_t bytesPerPixel);
 			void Copy(const Math::Vec2u& dstPos, Surface* surface, const Math::Vec2u& srcSize);
 			const Math::Vec2u& Size();
 			const Math::Vec2u& Quad();
