+++ source/Shared/Graphics/OpenGL3/TextureImplOpenGL3.hpp
@@ -14,13 +14,13 @@ namespace LDL
 		class TextureImplOpenGL3 : public TextureImpl
 		{
 		public:
-			TextureImplOpenGL3(RenderContextImpl* renderContextImpl, const Math::Vec2u & size, uint8_t* pixels, size_t bytesPerPixel);
-			TextureImplOpenGL3(RenderContextImpl* renderContextImpl, const Math::Vec2u& size, size_t bytesPerPixel);
+			TextureImplOpenGL3(RenderContextImpl* renderContextImpl, const Math::Vec2u & size, uint8_t* pixels, uint8_t bytesPerPixel);
+			TextureImplOpenGL3(RenderContextImpl* renderContextImpl, const Math::Vec2u& size, uint8_t bytesPerPixel);
 			~TextureImplOpenGL3();
 			const Math::Vec2u& Size();
 			const Math::Vec2u& Quad();
 			size_t Id();
-			void Copy(const Math::Vec2u& dstPos, const Math::Vec2u& srcSize, uint8_t* pixels, size_t bytesPerPixel);
+			void Copy(const Math::Vec2u& dstPos, const Math::Vec2u& srcSize, uint8_t* pixels, uint8_t bytesPerPixel);
 			void Copy(const Math::Vec2u& dstPos, Surface* surface, const Math::Vec2u& srcSize);
 		private:
 			RenderContextImpl* _RenderContextImpl;
