+++ source/Shared/Graphics/Software/TextureImplSoftware.hpp
@@ -15,9 +15,9 @@ namespace LDL
 		{
 		public:
 			TextureImplSoftware(RenderContextImpl* renderContextImpl, const LDL::Math::Vec2u & size, uint8_t* pixels, uint8_t bytesPerPixel);
-			TextureImplSoftware(RenderContextImpl* renderContextImpl, const Math::Vec2u& size, size_t bytesPerPixel);
+			TextureImplSoftware(RenderContextImpl* renderContextImpl, const Math::Vec2u& size, uint8_t bytesPerPixel);
 			~TextureImplSoftware();
-			void Copy(const Math::Vec2u& dstPos, const Math::Vec2u& srcSize, uint8_t* pixels, size_t bytesPerPixel);
+			void Copy(const Math::Vec2u& dstPos, const Math::Vec2u& srcSize, uint8_t* pixels, uint8_t bytesPerPixel);
 			void Copy(const Math::Vec2u& dstPos, Surface* surface, const Math::Vec2u& srcSize);
 			const Math::Vec2u& Size();
 			const Math::Vec2u& Quad();
